+++ packages/blocks/core/js/wordpress/column/test/functionality.blocks-1.e2e.cy.js
@@ -274,7 +274,24 @@ describe('Column Block', () => {
 			});
 
 		//
-		// 2. Assert inner blocks selectors in front end
+		// 2. Check settings tab
+		//
+		setParentBlock();
+		cy.getByDataTest('settings-tab').click();
+
+		// layout settings should be hidden
+		cy.get('.block-editor-block-inspector').within(() => {
+			cy.get('.components-panel__body-title button')
+				.contains('Layout')
+				.should('be.visible');
+
+			cy.get('.components-tools-panel-header')
+				.contains('Settings')
+				.should('not.be.visible');
+		});
+
+		//
+		// 3. Assert inner blocks selectors in front end
 		//
 		savePage();
 		redirectToFrontPage();